DFPL-2358: tidy up
