+++ playwright-e2e/pages/respondent-details.ts
@@ -1,14 +1,11 @@
 import {type Page, type Locator, expect} from "@playwright/test";
-
-export class RespondentDetails {
-
+import { BasePage } from "./base-page";
+export class RespondentDetails extends BasePage{
   readonly page: Page;
   readonly respondentDetailsHeading: Locator;
   readonly firstName: Locator;
   readonly lastName: Locator;
-  readonly dobDay: Locator; //DOB (date of birth)
-  readonly dobMonth: Locator;
-  readonly dobYear: Locator;
+  
   readonly gender: Locator;
   readonly currentAddress: Locator;
   readonly telephone: Locator;
@@ -18,18 +15,14 @@ export class RespondentDetails {
   readonly litigationCapacity: Locator; //ie Ability to take part in proceedings
   readonly litigationCapacityReason: Locator;
   readonly legalRepresentation: Locator;
-  readonly continue: Locator;
-  readonly saveAndContinue: Locator;
   readonly addressNotKnownReason: Locator;
 
   public constructor(page: Page) {
+    super(page);
     this.page = page;
     this.respondentDetailsHeading = page.getByRole("heading", { name: 'Respondents\' details' });
     this.firstName = page.getByLabel('*First name (Optional)');
     this.lastName = page.getByLabel('*Last name (Optional)');
-    this.dobDay = page.getByLabel('Day');
-    this.dobMonth = page.getByLabel('Month');
-    this.dobYear = page.getByLabel('Year');
     this.gender = page.getByLabel('Gender (Optional)');
     this.currentAddress = page.getByRole('group', { name: '*Current address known?' });
     this.addressNotKnownReason = page.getByLabel('*Reason the address is not known');
@@ -40,37 +33,25 @@ export class RespondentDetails {
     this.litigationCapacity = page.getByRole('group', { name: 'Do you believe this person will have problems with litigation capacity (understanding what\'s happening in the case)? (Optional)' });
     this.litigationCapacityReason = page.getByLabel('Give details, including assessment outcomes and referrals to health services (Optional)');
     this.legalRepresentation = page.getByRole('group', { name: '*Do they have legal representation? (Optional)' });
-    this.continue = page.getByRole('button', { name: 'Continue' });
-    this.saveAndContinue = page.getByRole('button', { name: 'Save and continue' });
   }
 
   async respondentDetailsNeeded() {
     await expect(this.respondentDetailsHeading).toBeVisible();
-    await this.firstName.click();
     await this.firstName.fill('John');
-    await this.lastName.click();
     await this.lastName.fill('Smith');
-    await this.dobDay.click();
-    await this.dobDay.fill('10');
-    await this.dobMonth.click();
-    await this.dobMonth.fill('11');
-    await this.dobYear.click();
-    await this.dobYear.fill('2001');
+    await this.dobFillOut("1","12",(new Date().getUTCFullYear()-20).toString());
     await this.gender.click(); //not sure if click needed
     await this.gender.selectOption('1: Male');
     await this.currentAddress.getByLabel('No').check();
     await this.addressNotKnownReason.selectOption('2: Person deceased');
     await this.telephone.fill('01234567890');
-    await this.relationToChild.click();
     await this.relationToChild.fill('aunt');
     await this.relationToChildContact.getByLabel('Yes').check();
-    await this.relationToChildContactReason.click();
     await this.relationToChildContactReason.fill('this is the reason');
     await this.litigationCapacity.getByLabel('Yes').check();
-    await this.litigationCapacityReason.click();
     await this.litigationCapacityReason.fill('these are the details');
     await this.legalRepresentation.getByLabel('No').check();
-    await this.continue.click();
-    await this.saveAndContinue.click();
+    await this.clickContinue();
+    await this.checkYourAnsAndSubmit();
   }
 }
